python/requirements.txt
pytest
jpype1
stubgenj